Major command-line app refactor
