+++ _cli/sparrow_cli/commands/info.py
@@ -1,7 +1,6 @@
 import click
 from ..config import SparrowConfig
 from ..meta import __version__
-from json import load
 
 
 def sparrow_core_version(cfg):
